pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.8.1",
    "cryptography>=44.0.2",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
